% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tutorial.R
\name{run_tutorial}
\alias{run_tutorial}
\title{Cost-Effectiveness Simulation Modelling Tutorial}
\usage{
run_tutorial(name = "full-model")
}
\arguments{
\item{name}{Name of the tutorial to run. Currently only the 'full-model'
tutorial (using our previously defined model structures to implement a
modelled cost-effectiveness analysis) is available.}
}
\description{
Run the 'Cost-Effectiveness Analysis' tutorial on simulation modelling.
}
